Split heap_wrapper_apis
